pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "singlestore-haystack"
dynamic = ["version"]
description = 'Haystack 2.x Document Store for SingleStore'
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
keywords = []
authors = [
    { name = "SingleStore", email = "integration@singlestore.com" },
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "haystack-ai>=2.11.0",
    "typing_extensions",
    "singlestoredb>=1.16.6"
]

[project.urls]
Documentation = "https://github.com/singlestore-labs/singlestore-haystack#readme"
Issues = "https://github.com/singlestore-labs/singlestore-haystack/issues"
Source = "https://github.com/singlestore-labs/singlestore-haystack"

[tool.hatch.build.targets.wheel]
packages = ["src/haystack_integrations"]

[tool.hatch.version]
source = "vcs"
tag-pattern = 'v(?P<version>.*)'

[tool.hatch.envs.default]
dependencies = [
    "ruff",
    "pytest",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
fmt = "ruff check --fix {args} && ruff format {args}"
fmt-check = "ruff check {args} && ruff format --check {args}"

[tool.hatch.envs.test]
dependencies = [
    "pytest",
    "pytest-asyncio",
    "pytest-rerunfailures",
    "mypy",
    "pip"
]

[tool.hatch.envs.test.scripts]
all = 'pytest {args:tests}'
types = "mypy -p haystack_integrations.document_stores.singlestore_haystack -p haystack_integrations.components.retrievers.singlestore_haystack {args}"

[tool.mypy]
install_types = true
non_interactive = true
check_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    # singlestoredb does not provide types
    "singlestoredb.*",
]
ignore_missing_imports = true


[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Allow boolean positional values in function calls, like `dict.get(... True)`
    "FBT003",
    # Ignore checks for possible passwords
    "S105",
    "S106",
    "S107",
    # Ignore complexity
    "C901",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    # ignore function-call-in-default-argument
    "B008",
    # ignore assertions
    "S101",
]
unfixable = [
    # Don't touch unused imports
    "F401",
]

[tool.ruff.lint.isort]
known-first-party = ["haystack_integrations"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
# examples can contain "print" commands
"examples/**/*" = ["T201"]
